update indexedDBmock tests to run with mocha

+++ test/test-helper.js
@@ -1,7 +1,8 @@
 require('../src/idb-polyfill').polyfill()
 
-before(function() { localStorage.clear() })
-after(function() { localStorage.clear() })
+if (typeof localStorage !== 'undefined') {
+  before(function() { localStorage.clear() })
+  after(function() { localStorage.clear() })
+}
 
-mocha.setup({ timeout: 10000 })
 module.exports = global.pIndexedDB || global.indexedDB
